ajout du dao/service/controller pour les messages

+++ src/main/java/kaboni/libertchan/dao/BannedUserJpaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import kaboni.libertchan.models.BannedUser;
 
 @Repository
-public interface BannedUserJpaRepository extends JpaRepository<BannedUser, Integer> {
+public interface BannedUserJpaRepository extends JpaRepository<BannedUser, Long> {
 		public Optional<BannedUser> findByIp(String ip);
 		public Optional<BannedUser> findById(Long id);
 }
